Fix focus issue and add flexibility `enableUserSelectHack` via props from GridItem (#3)
**Target problem**:
Whenever there is an input and a draggable list and the user types something, the focus is lost from the input field because the `enableUserSelectHack` is removing some styles

**Solution description**:
This flag is now set as `false` when using the `GridLayout` component. This will fix performance issues and also allow this flag control via props.